+++ .env.conduit-user.example
@@ -1,3 +1,7 @@
 CONDUIT_SERVICE_URI=
 CONDUIT_SERVICE_API_KEY=
 CONDUIT_SERVICE_OBJECT_KEY=
+USER_FIRST_NAME=
+USER_LAST_NAME=
+USER_EMAIL=
+USER_PASSWORD=
