+++ .gitignore
@@ -56,7 +56,7 @@ typings/
 
 # dotenv environment variables file
 .env
-.env.conduit
+.env.conduit-user
 
 # next.js build output
 .next
